clean out uploaded files before testing
