--- wsb-pilot-workspace/s3api/get-object/sample3.txt
@@ -1,1 +1,0 @@
-sample3
